Convert Icon to TS
